implements fetching previous object on user path id input
previously it just used the initial text/image, but now when a user types for example f1, the last direction was 1 which was right. and the last path id was f. the game downloads both path f1, and path f. so when it renders path id f1 it can render the text/image of path f. closes #31