changes bundle to use contenthash
i misunderstood the effect when I first changed it. I want contenthash because it uses a specific hash for EACH file, instead of the whole code combined. when using contenthash it will create a seperate hash for the npm packages bundle, which will be the same throughout multiple deployments because the packages have not changed. this way, whenever I make a new deployment, users will only have to redownload MY code, and not the entire npm packages Im using